Separate 'create new file/folder' and 'upload files/folder' actions into two dropdowns
Extract create/upload from appBar into new component
