+++ packages/web-app-files/src/components/AppBar/CreateAndUpload.vue
@@ -1,0 +1,567 @@
+<template>
+  <div class="oc-flex" style="gap: 15px">
+    <template v-if="createFileActionsAvailable">
+      <oc-button
+        id="new-file-menu-btn"
+        key="new-file-menu-btn-enabled"
+        v-oc-tooltip="newButtonTooltip"
+        :aria-label="newButtonAriaLabel"
+        variation="inverse"
+        :disabled="uploadOrFileCreationBlocked"
+        class="oc-background-primary-gradient"
+      >
+        <oc-icon name="add" />
+        <translate>New</translate>
+      </oc-button>
+      <oc-drop
+        drop-id="new-file-menu-drop"
+        toggle="#new-file-menu-btn"
+        mode="click"
+        close-on-click
+        class="oc-width-auto"
+        padding-size="small"
+      >
+        <oc-list id="create-list">
+          <li class="create-list-folder oc-py-xs">
+            <oc-button id="new-folder-btn" appearance="raw" @click="showCreateResourceModal">
+              <oc-resource-icon :resource="{ isFolder: true, extension: '' }" />
+              <translate>Folder</translate>
+            </oc-button>
+          </li>
+          <li
+            v-for="(newFileHandler, key) in newFileHandlersForRoute"
+            :key="key"
+            class="create-list-file oc-py-xs"
+          >
+            <oc-button
+              appearance="raw"
+              :class="['new-file-btn-' + newFileHandler.ext]"
+              @click="showCreateResourceModal(false, newFileHandler.ext, newFileHandler.action)"
+            >
+              <oc-resource-icon :resource="{ type: 'file', extension: newFileHandler.ext }" />
+              <span>{{ newFileHandler.menuTitle($gettext) }}</span>
+            </oc-button>
+          </li>
+          <template v-if="mimetypesAllowedForCreation">
+            <li
+              v-for="(mimetype, key) in mimetypesAllowedForCreation"
+              :key="key"
+              class="create-list-file oc-py-xs"
+            >
+              <oc-button
+                appearance="raw"
+                @click="showCreateResourceModal(false, mimetype.ext, false, true)"
+              >
+                <oc-resource-icon :resource="{ type: 'file', extension: mimetype.ext }" />
+                <translate :translate-params="{ name: mimetype.name }">%{name}</translate>
+              </oc-button>
+            </li>
+          </template>
+        </oc-list>
+      </oc-drop>
+    </template>
+    <template v-else>
+      <oc-button
+        id="new-folder-btn"
+        v-oc-tooltip="newButtonAriaLabel"
+        variation="inverse"
+        :aria-label="newButtonAriaLabel"
+        :disabled="uploadOrFileCreationBlocked"
+        class="oc-background-primary-gradient"
+        @click="showCreateResourceModal"
+      >
+        <oc-icon name="resource-type-folder" />
+        <translate style="display: ruby">New folder</translate>
+      </oc-button>
+    </template>
+    <oc-button
+      id="upload-menu-btn"
+      key="upload-menu-btn-enabled"
+      v-oc-tooltip="uploadButtonTooltip"
+      :aria-label="uploadButtonAriaLabel"
+      :disabled="uploadOrFileCreationBlocked"
+    >
+      <oc-icon name="upload" fill-type="line" />
+      <translate>Upload</translate>
+    </oc-button>
+    <oc-drop
+      drop-id="upload-menu-drop"
+      toggle="#upload-menu-btn"
+      mode="click"
+      class="oc-width-auto"
+      close-on-click
+      padding-size="small"
+    >
+      <oc-list id="upload-list">
+        <li class="oc-p-s">
+          <folder-upload
+            v-if="!isIE11()"
+            :root-path="currentPath"
+            :path="currentPath"
+            :headers="headers"
+            @success="uploadSuccess"
+            @error="uploadError"
+            @progress="uploadProgress"
+          />
+        </li>
+        <li class="oc-p-s">
+          <file-upload
+            :path="currentPath"
+            :headers="headers"
+            @success="uploadSuccess"
+            @error="uploadError"
+            @progress="uploadProgress"
+          />
+        </li>
+      </oc-list>
+    </oc-drop>
+  </div>
+</template>
+
+<script>
+import { mapActions, mapGetters, mapState, mapMutations } from 'vuex'
+import pathUtil from 'path'
+
+import Mixins from '../../mixins'
+import MixinFileActions, { EDITOR_MODE_CREATE } from '../../mixins/fileActions'
+import { buildResource, buildWebDavFilesPath, buildWebDavSpacesPath } from '../../helpers/resources'
+import { isLocationActive, isLocationPublicActive, isLocationSpacesActive } from '../../router'
+import { useActiveLocation } from '../../composables'
+
+import { DavProperties, DavProperty } from 'web-pkg/src/constants'
+
+import FileUpload from './Upload/FileUpload.vue'
+import FolderUpload from './Upload/FolderUpload.vue'
+
+export default {
+  components: {
+    FileUpload,
+    FolderUpload
+  },
+  mixins: [Mixins, MixinFileActions],
+  props: {
+    canUpload: { type: Boolean, default: false },
+    currentPath: { type: String, default: null },
+    hasFreeSpace: { type: Boolean, default: false },
+    headers: { type: Object, default: null }
+  },
+  setup() {
+    return {
+      isPersonalLocation: useActiveLocation(isLocationSpacesActive, 'files-spaces-personal-home'),
+      isPublicLocation: useActiveLocation(isLocationPublicActive, 'files-public-files'),
+      isSpacesProjectsLocation: useActiveLocation(isLocationSpacesActive, 'files-spaces-projects'),
+      isSpacesProjectLocation: useActiveLocation(isLocationSpacesActive, 'files-spaces-project')
+    }
+  },
+  computed: {
+    ...mapGetters([
+      'getToken',
+      'capabilities',
+      'configuration',
+      'newFileHandlers',
+      'quota',
+      'user'
+    ]),
+    ...mapGetters('Files', ['files', 'currentFolder', 'selectedFiles', 'publicLinkPassword']),
+    ...mapState('Files', ['areHiddenFilesShown']),
+
+    mimetypesAllowedForCreation() {
+      // we can't use `mapGetters` here because the External app doesn't exist in all deployments
+      const mimeTypes = this.$store.getters['External/mimeTypes']
+      if (!mimeTypes) {
+        return []
+      }
+      return mimeTypes.filter((mimetype) => mimetype.allow_creation) || []
+    },
+    createFileActionsAvailable() {
+      return this.newFileHandlersForRoute.length > 0 || this.mimetypesAllowedForCreation.length > 0
+    },
+    newButtonTooltip() {
+      if (!this.canUpload) {
+        return this.$gettext('You have no permission to create new files!')
+      }
+      if (!this.hasFreeSpace) {
+        return this.$gettext('You have not enough space left to create new files!')
+      }
+      return null
+    },
+    newButtonAriaLabel() {
+      const tooltip = this.newButtonTooltip
+      if (tooltip) {
+        return tooltip
+      }
+      if (!this.createFileActionsAvailable) {
+        return this.$gettext('Create a new folder')
+      }
+      return this.$gettext('Create new files or folders')
+    },
+    uploadButtonTooltip() {
+      if (!this.canUpload) {
+        return this.$gettext('You have no permission to upload!')
+      }
+      if (!this.hasFreeSpace) {
+        return this.$gettext('You have not enough space left to upload!')
+      }
+      return null
+    },
+    uploadButtonAriaLabel() {
+      const tooltip = this.uploadButtonTooltip
+      if (tooltip) {
+        return tooltip
+      }
+      return this.$gettext('Upload files or folders')
+    },
+
+    uploadOrFileCreationBlocked() {
+      return !this.canUpload || !this.hasFreeSpace
+    },
+
+    newFileHandlersForRoute() {
+      return this.newFileHandlers.filter(({ routes = [] }) =>
+        isLocationActive(this.$router, ...routes.map((name) => ({ name })))
+      )
+    }
+  },
+  methods: {
+    ...mapActions('Files', ['loadIndicators']),
+    ...mapActions(['openFile', 'showMessage', 'createModal', 'setModalInputErrorMessage']),
+    ...mapMutations('Files', ['UPSERT_RESOURCE']),
+
+    uploadSuccess() {
+      this.$emit('success')
+    },
+    uploadError() {
+      this.$emit('error')
+    },
+    uploadProgress() {
+      this.$emit('progress')
+    },
+
+    showCreateResourceModal(
+      isFolder = true,
+      ext = 'txt',
+      openAction = null,
+      addAppProviderFile = false
+    ) {
+      const defaultName = isFolder
+        ? this.$gettext('New folder')
+        : this.$gettext('New file') + '.' + ext
+      const checkInputValue = (value) => {
+        this.setModalInputErrorMessage(
+          isFolder ? this.checkNewFolderName(value) : this.checkNewFileName(value)
+        )
+      }
+
+      // Sets action to be executed after creation of the file
+      if (!isFolder) {
+        this.newFileAction = openAction
+      }
+
+      const modal = {
+        variation: 'passive',
+        title: isFolder ? this.$gettext('Create a new folder') : this.$gettext('Create a new file'),
+        cancelText: this.$gettext('Cancel'),
+        confirmText: this.$gettext('Create'),
+        hasInput: true,
+        inputValue: defaultName,
+        inputLabel: isFolder ? this.$gettext('Folder name') : this.$gettext('File name'),
+        inputError: isFolder
+          ? this.checkNewFolderName(defaultName)
+          : this.checkNewFileName(defaultName),
+        onCancel: this.hideModal,
+        onConfirm: isFolder
+          ? this.addNewFolder
+          : addAppProviderFile
+          ? this.addAppProviderFile
+          : this.addNewFile,
+        onInput: checkInputValue
+      }
+
+      this.createModal(modal)
+    },
+
+    async addNewFolder(folderName) {
+      if (folderName === '') {
+        return
+      }
+
+      this.fileFolderCreationLoading = true
+
+      try {
+        let path = pathUtil.join(this.currentPath, folderName)
+        let resource
+
+        if (this.isPersonalLocation) {
+          path = buildWebDavFilesPath(this.user.id, path)
+          await this.$client.files.createFolder(path)
+          resource = await this.$client.files.fileInfo(path, DavProperties.Default)
+        } else if (this.isSpacesProjectLocation) {
+          path = buildWebDavSpacesPath(this.$route.params.spaceId, path)
+          await this.$client.files.createFolder(path)
+          resource = await this.$client.files.fileInfo(path, DavProperties.Default)
+        } else {
+          await this.$client.publicFiles.createFolder(path, null, this.publicLinkPassword)
+          resource = await this.$client.publicFiles.getFileInfo(
+            path,
+            this.publicLinkPassword,
+            DavProperties.PublicLink
+          )
+        }
+        resource = buildResource(resource)
+
+        this.UPSERT_RESOURCE(resource)
+        this.hideModal()
+
+        if (this.isPersonalLocation) {
+          this.loadIndicators({
+            client: this.$client,
+            currentFolder: this.currentFolder.path
+          })
+        }
+
+        this.showMessage({
+          title: this.$gettextInterpolate(
+            this.$gettext('"%{folderName}" was created successfully'),
+            {
+              folderName
+            }
+          )
+        })
+      } catch (error) {
+        console.error(error)
+        this.showMessage({
+          title: this.$gettext('Failed to create folder'),
+          status: 'danger'
+        })
+      }
+
+      this.fileFolderCreationLoading = false
+    },
+
+    checkNewFolderName(folderName) {
+      if (folderName === '') {
+        return this.$gettext('Folder name cannot be empty')
+      }
+
+      if (/[/]/.test(folderName)) {
+        return this.$gettext('Folder name cannot contain "/"')
+      }
+
+      if (folderName === '.') {
+        return this.$gettext('Folder name cannot be equal to "."')
+      }
+
+      if (folderName === '..') {
+        return this.$gettext('Folder name cannot be equal to ".."')
+      }
+
+      if (/\s+$/.test(folderName)) {
+        return this.$gettext('Folder name cannot end with whitespace')
+      }
+
+      const exists = this.files.find((file) => file.name === folderName)
+
+      if (exists) {
+        const translated = this.$gettext('%{name} already exists')
+        return this.$gettextInterpolate(translated, { name: folderName }, true)
+      }
+
+      return null
+    },
+
+    async addNewFile(fileName) {
+      if (fileName === '') {
+        return
+      }
+
+      this.fileFolderCreationLoading = true
+
+      try {
+        let resource
+        let path = pathUtil.join(this.currentPath, fileName)
+
+        if (this.isPersonalLocation) {
+          path = buildWebDavFilesPath(this.user.id, path)
+          await this.$client.files.putFileContents(path, '')
+          resource = await this.$client.files.fileInfo(path, DavProperties.Default)
+        } else if (this.isSpacesProjectLocation) {
+          path = buildWebDavSpacesPath(this.$route.params.spaceId, path)
+          await this.$client.files.putFileContents(path, '')
+          resource = await this.$client.files.fileInfo(path, DavProperties.Default)
+        } else {
+          await this.$client.publicFiles.putFileContents('', path, this.publicLinkPassword, '')
+          resource = await this.$client.publicFiles.getFileInfo(
+            path,
+            this.publicLinkPassword,
+            DavProperties.PublicLink
+          )
+        }
+
+        if (this.newFileAction) {
+          const fileId = resource.fileInfo[DavProperty.FileId]
+
+          this.$_fileActions_openEditor(this.newFileAction, path, fileId, EDITOR_MODE_CREATE)
+          this.hideModal()
+
+          return
+        }
+
+        resource = buildResource(resource)
+
+        this.UPSERT_RESOURCE(resource)
+        this.hideModal()
+
+        if (this.isPersonalLocation) {
+          this.loadIndicators({
+            client: this.$client,
+            currentFolder: this.currentFolder.path
+          })
+        }
+
+        this.showMessage({
+          title: this.$gettextInterpolate(this.$gettext('"%{fileName}" was created successfully'), {
+            fileName
+          })
+        })
+      } catch (error) {
+        console.error(error)
+        this.showMessage({
+          title: this.$gettext('Failed to create file'),
+          status: 'danger'
+        })
+      }
+
+      this.fileFolderCreationLoading = false
+    },
+    async addAppProviderFile(fileName) {
+      // FIXME: this belongs in web-app-external, but the app provider handles file creation differently than other editor extensions. Needs more refactoring.
+      if (fileName === '') {
+        return
+      }
+      try {
+        const parent = this.currentFolder.fileId
+        const publicToken = (this.$router.currentRoute.params.item || '').split('/')[0]
+
+        const configUrl = this.configuration.server
+        const appNewUrl = this.capabilities.files.app_providers[0].new_url.replace(/^\/+/, '')
+        const url =
+          configUrl +
+          appNewUrl +
+          `?parent_container_id=${parent}&filename=${encodeURIComponent(fileName)}`
+
+        const headers = {
+          'X-Requested-With': 'XMLHttpRequest',
+          ...(this.isPublicLocation &&
+            publicToken && {
+              'public-token': publicToken
+            }),
+          ...(this.isPublicLocation &&
+            this.publicLinkPassword && {
+              Authorization:
+                'Basic ' +
+                Buffer.from(['public', this.publicLinkPassword].join(':')).toString('base64')
+            }),
+          ...(this.getToken && {
+            Authorization: 'Bearer ' + this.getToken
+          })
+        }
+
+        const response = await fetch(url, {
+          method: 'POST',
+          headers
+        })
+
+        if (response.status !== 200) {
+          throw new Error(`An error has occurred: ${response.status}`)
+        }
+
+        let resource
+        let path = pathUtil.join(this.currentPath, fileName)
+        if (this.isPersonalLocation) {
+          path = buildWebDavFilesPath(this.user.id, path)
+          resource = await this.$client.files.fileInfo(path, DavProperties.Default)
+        } else if (this.isSpacesProjectLocation) {
+          path = buildWebDavSpacesPath(this.$route.params.spaceId, path)
+          resource = await this.$client.files.fileInfo(path, DavProperties.Default)
+        } else {
+          resource = await this.$client.publicFiles.getFileInfo(
+            path,
+            this.publicLinkPassword,
+            DavProperties.PublicLink
+          )
+        }
+        resource = buildResource(resource)
+        this.$_fileActions_triggerDefaultAction(resource)
+        this.UPSERT_RESOURCE(resource)
+        this.hideModal()
+
+        if (this.isPersonalLocation) {
+          this.loadIndicators({
+            client: this.$client,
+            currentFolder: this.currentFolder.path
+          })
+        }
+        this.showMessage({
+          title: this.$gettextInterpolate(this.$gettext('"%{fileName}" was created successfully'), {
+            fileName
+          })
+        })
+      } catch (error) {
+        console.error(error)
+        this.showMessage({
+          title: this.$gettext('Failed to create file'),
+          status: 'danger'
+        })
+      }
+    },
+    checkNewFileName(fileName) {
+      if (fileName === '') {
+        return this.$gettext('File name cannot be empty')
+      }
+
+      if (/[/]/.test(fileName)) {
+        return this.$gettext('File name cannot contain "/"')
+      }
+
+      if (fileName === '.') {
+        return this.$gettext('File name cannot be equal to "."')
+      }
+
+      if (fileName === '..') {
+        return this.$gettext('File name cannot be equal to ".."')
+      }
+
+      if (/\s+$/.test(fileName)) {
+        return this.$gettext('File name cannot end with whitespace')
+      }
+
+      const exists = this.files.find((file) => file.name === fileName)
+
+      if (exists) {
+        const translated = this.$gettext('%{name} already exists')
+        return this.$gettextInterpolate(translated, { name: fileName }, true)
+      }
+
+      return null
+    }
+  }
+}
+</script>
+<style lang="scss" scoped>
+#create-list {
+  li {
+    border: 1px solid transparent;
+    button {
+      gap: 10px;
+      justify-content: left;
+      width: 100%;
+    }
+  }
+  .create-list-folder {
+    border-bottom: 1px solid var(--oc-color-border);
+  }
+  .create-list-file button {
+    margin: 2px 0;
+  }
+}
+</style>
