+++ packages/web-app-files/tests/unit/components/AppBar/Upload/FolderUpload.spec.js
@@ -44,6 +44,7 @@ describe('FolderUpload Component', () => {
           ...mountOptions,
           stubs: {
             'oc-icon': true,
+            'oc-resource-icon': true,
             'oc-button': true
           }
         })
