+++ tests/acceptance/pageObjects/personalPage.js
@@ -125,8 +125,8 @@ module.exports = {
       return this
     },
     selectFileForUpload: function (filePath) {
-      return this.waitForElementVisible('@newFileMenuButton')
-        .click('@newFileMenuButton')
+      return this.waitForElementVisible('@uploadFilesButton')
+        .click('@uploadFilesButton')
         .waitForElementVisible('@fileUploadButton')
         .setValue('@fileUploadInput', filePath)
     },
@@ -313,6 +313,9 @@ module.exports = {
   elements: {
     searchInput: {
       selector: 'input.oc-search-input'
+    },
+    uploadFilesButton: {
+      selector: '#upload-menu-btn'
     },
     newFileMenuButtonAnyState: {
       selector: '#new-file-menu-btn'
